+++ packages/plugins/Attributions/src/components/Attributions.vue
@@ -3,12 +3,7 @@
     v-if="renderType === 'independent'"
     class="polar-plugin-attributions-wrapper"
   >
-    <AttributionContent
-      v-if="windowIsOpen"
-      class="mr-2"
-      :max-width="maxWidth"
-      :width="windowWidth"
-    ></AttributionContent>
+    <AttributionContent v-if="windowIsOpen" class="mr-2"></AttributionContent>
     <AttributionButton></AttributionButton>
   </div>
   <AttributionContent v-else></AttributionContent>
@@ -27,17 +22,7 @@ export default Vue.extend({
     AttributionContent,
   },
   computed: {
-    ...mapGetters(['clientWidth', 'hasSmallWidth', 'hasWindowSize']),
-    ...mapGetters('plugin/attributions', [
-      'renderType',
-      'windowIsOpen',
-      'windowWidth',
-    ]),
-    maxWidth() {
-      return this.hasWindowSize && this.hasSmallWidth
-        ? this.clientWidth * 0.85
-        : 1080
-    },
+    ...mapGetters('plugin/attributions', ['renderType', 'windowIsOpen']),
   },
 })
 </script>